ENH: added some double support

+++ graphics/vtkFollower.cxx
@@ -64,7 +64,7 @@ vtkFollower::~vtkFollower()
 // Copy the follower's composite 4x4 matrix into the matrix provided.
 void vtkFollower::GetMatrix(vtkMatrix4x4 *result)
 {
-  float *pos;
+  double *pos;
   vtkMatrix4x4 *matrix = vtkMatrix4x4::New();
 
   this->GetOrientation();
@@ -92,13 +92,13 @@ void vtkFollower::GetMatrix(vtkMatrix4x4 *result)
   // add the rotation to follow the camera
   if (this->Camera)
     {
-    float distance, distance_old;
-    float *vup;
-    float twist = 0;
-    float v1[3], v2[3], y_axis[3];
+    double distance, distance_old;
+    double *vup;
+    double twist = 0;
+    double v1[3], v2[3], y_axis[3];
     double theta, dot, mag;
     double cosang;
-    float vn[3];
+    double vn[3];
 
     // calc the direction
     pos = this->Camera->GetPosition();
